refactor: firebase (#29)
* refactor: firebase

* fix: eslint

* fix: alias

* fix: eslint
