+++ package.json
@@ -7,7 +7,8 @@
     "dev": "env-cmd -f config/.env.local next",
     "build": "next build",
     "start": "next start",
-    "lint": "eslint --fix",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix",
     "format": "prettier --write . --config config/.prettierrc.json --ignore-path config/.prettierignore",
     "prepare": "husky install",
     "pre-commit": "lint-staged -c config/.lintstagedrc.json && npm run lint && npm run type-check",
