+++ src/apis/client/users.ts
@@ -1,6 +1,6 @@
 import { FIRESTORE } from "@/constants";
 import { User } from "@/types";
-import { firestore } from "@/utils/firebase/client";
+import { firestore } from "src/firebase/client";
 
 export const fetchUser = async (uid: string) =>
   firestore.getDoc<User>(FIRESTORE.COLLECTION.USERS, uid);
