+++ src/firebase/admin/auth/index.ts
@@ -1,0 +1,1 @@
+export { getAuth } from "firebase-admin/auth";
