+++ src/firebase/admin/firestore/index.ts
@@ -1,0 +1,1 @@
+export { getFirestore } from "firebase-admin/firestore";
