+++ src/firebase/client/auth/index.ts
@@ -1,0 +1,1 @@
+export { getAuth } from "firebase/auth";
