+++ src/firebase/client/index.ts
@@ -1,0 +1,5 @@
+import { app } from "./app";
+
+const { auth, firestore } = app;
+
+export { auth, firestore, app as firebase };
