+++ src/repositories/auth.ts
@@ -1,11 +1,9 @@
 import {
   createUserWithEmailAndPassword,
   signInWithRedirect,
+  GoogleAuthProvider,
 } from "firebase/auth";
-import { GoogleAuthProvider } from "firebase/auth";
-import { auth } from "~/firebase/client";
-
-export const googleProvider = new GoogleAuthProvider();
+import { auth } from "src/firebase/client";
 
 export const createAccountWithEmailAndPassword = async (
   email: string,
@@ -16,6 +14,7 @@ export const createAccountWithEmailAndPassword = async (
  * @see https://firebase.google.com/docs/auth/web/google-signin
  */
 export const signInWithGoogle = async () => {
+  const googleProvider = new GoogleAuthProvider();
   return signInWithRedirect(auth, googleProvider);
   // This gives you a Google Access Token. You can use it to access Google APIs.
   // const credential = GoogleAuthProvider.credentialFromResult(result);
