--- src/utils/firebase/admin/index.ts
@@ -1,3 +1,0 @@
-import app from "./app";
-
-export const { auth, firestore } = app;
