--- src/utils/firebase/client/app.ts
@@ -1,40 +1,0 @@
-import { getAnalytics } from "firebase/analytics";
-import { initializeApp, FirebaseApp } from "firebase/app";
-import { getAuth } from "firebase/auth";
-import { getFirestore } from "firebase/firestore";
-
-const clientCredentials = {
-  apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY,
-  authDomain: process.env.NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN,
-  databaseURL: process.env.NEXT_PUBLIC_FIREBASE_DATABASE_URL,
-  projectId: process.env.NEXT_PUBLIC_FIREBASE_PROJECT_ID,
-  storageBucket: process.env.NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID,
-  appId: process.env.NEXT_PUBLIC_FIREBASE_APP_ID,
-  measurementId: process.env.NEXT_PUBLIC_FIREBASE_MEASUREMENT_ID,
-};
-
-class App {
-  private _app: FirebaseApp;
-
-  constructor() {
-    this._app = initializeApp(clientCredentials);
-    // Check that `window` is in scope for the analytics module!
-    if (typeof window !== "undefined") {
-      // Enable analytics. https://firebase.google.com/docs/analytics/get-started
-      if ("measurementId" in clientCredentials) {
-        getAnalytics();
-      }
-    }
-  }
-
-  get auth() {
-    return getAuth(this._app);
-  }
-
-  get firestore() {
-    return getFirestore(this._app);
-  }
-}
-
-export default new App();
