--- src/utils/firebase/client/index.ts
@@ -1,6 +1,0 @@
-import app from "./app";
-import * as firestore from "./firestore";
-
-const { auth } = app;
-
-export { auth, firestore };
